refactor: restructure authentication routes by consolidating into a modular format, removing deprecated files, and enhancing route management for improved clarity and maintainability
