+++ backend/routes/auth/auth/refresh.js
@@ -1,0 +1,9 @@
+import express from 'express';
+import { refresh } from '../../controllers/auth/refresh.js';
+
+const router = express.Router();
+
+// Refresh token endpoint
+router.post('/', refresh);
+
+export default router;
