+++ backend/routes/auth/auth/signup.js
@@ -1,0 +1,9 @@
+import express from 'express';
+import { signup } from '../../controllers/auth/signup.js';
+
+const router = express.Router();
+
+// User signup endpoint
+router.post('/', signup);
+
+export default router;
