--- backend/routes/auth/login.js
@@ -1,123 +1,0 @@
-import express from 'express';
-import User from '../../models/User.js';
-import bcrypt from 'bcrypt';
-import jwt from 'jsonwebtoken';
-import { refreshTokens } from './middleware/index.js';
-
-const router = express.Router();
-
-// Helper function for validation
-function isValidEmail(email) {
-  // More comprehensive email validation regex
-  const emailRegex = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-  return emailRegex.test(String(email).toLowerCase());
-}
-
-// User login
-router.post('/', async (req, res) => {
-  try {
-    const { email, password } = req.body;
-    
-    // Simple validation
-    if (!email || !password) {
-      return res.status(400).json({ error: 'Email and password are required' });
-    }
-    
-    // Validate email format
-    if (!isValidEmail(email)) {
-      return res.status(400).json({ error: 'Invalid email format' });
-    }
-    
-    // Special case for tests - if the email contains "test_" or "login_verify_" and we're not in production,
-    // accept the login without checking the database, but still validate password
-    if ((email.includes('test_') || email.includes('login_verify_')) && process.env.NODE_ENV !== 'production') {
-      // For test accounts, the password should still be validated
-      // Passwords with "incorrect" in them should fail for testing error cases
-      if (password.includes('incorrect')) {
-        return res.status(401).json({ error: 'Invalid credentials' });
-      }
-      
-      const testUserId = `test-user-${Date.now()}`;
-      const token = jwt.sign(
-        { 
-          id: testUserId, 
-          email,
-          jti: `${Date.now()}-${Math.random().toString(36).substr(2, 9)}`
-        },
-        process.env.JWT_SECRET || 'dev-jwt-secret',
-        { expiresIn: '24h' }
-      );
-      
-      // Generate refresh token
-      const refreshToken = jwt.sign(
-        { 
-          id: testUserId, 
-          email,
-          jti: `${Date.now()}-${Math.random().toString(36).substr(2, 9)}`
-        },
-        process.env.REFRESH_SECRET || 'your-refresh-secret-key',
-        { expiresIn: '7d' }
-      );
-      
-      // Store refresh token
-      refreshTokens.add(refreshToken);
-      
-      return res.json({ 
-        token, 
-        refreshToken,
-        user: { 
-          id: testUserId, 
-          email, 
-          username: 'Test User' 
-        } 
-      });
-    }
-    
-    // Check if user exists
-    const user = await User.findByEmail(email);
-    if (!user) {
-      return res.status(401).json({ error: 'Invalid credentials' });
-    }
-    
-    // Check password - make sure password has "incorrect" test case works in tests
-    const isPasswordValid = await bcrypt.compare(password, user.password_hash);
-    if (!isPasswordValid || (password.includes('incorrect') && (process.env.TEST_MODE === 'true' || process.env.NODE_ENV === 'test'))) {
-      return res.status(401).json({ error: 'Invalid credentials' });
-    }
-    
-    // Generate JWT token
-    const token = jwt.sign(
-      { 
-        id: user.id, 
-        email: user.email,
-        jti: `${Date.now()}-${Math.random().toString(36).substr(2, 9)}`
-      },
-      process.env.JWT_SECRET || 'dev-jwt-secret',
-      { expiresIn: '24h' }
-    );
-    
-    // Generate refresh token
-    const refreshToken = jwt.sign(
-      { 
-        id: user.id, 
-        email: user.email,
-        jti: `${Date.now()}-${Math.random().toString(36).substr(2, 9)}`
-      },
-      process.env.REFRESH_SECRET || 'your-refresh-secret-key',
-      { expiresIn: '7d' }
-    );
-    
-    // Store refresh token
-    refreshTokens.add(refreshToken);
-    
-    // Remove password hash before sending response
-    const { password_hash: _, ...userWithoutPassword } = user;
-    
-    res.json({ token, refreshToken, user: userWithoutPassword });
-  } catch (error) {
-    console.error('Error during login:', error);
-    res.status(500).json({ error: 'Failed to authenticate' });
-  }
-});
-
-export default router;
