--- backend/routes/auth/logout.js
@@ -1,51 +1,0 @@
-import express from 'express';
-import jwt from 'jsonwebtoken';
-import { authenticateToken, refreshTokens } from './middleware/index.js';
-
-const router = express.Router();
-
-// User logout
-router.post('/', authenticateToken, (req, res) => {
-  try {
-    const { refreshToken } = req.body;
-    
-    if (!refreshToken) {
-      return res.status(400).json({ error: 'Refresh token is required' });
-    }
-    
-    // Extract JWT Secret
-    const REFRESH_SECRET = process.env.REFRESH_SECRET || 'your-refresh-secret-key';
-    
-    // Always consider as test environment for consistent behavior with tests
-    const isTestMode = process.env.TEST_MODE === 'true' || process.env.NODE_ENV === 'test' || req.get('User-Agent')?.includes('node-superagent');
-    
-    // Verify refresh token is valid JWT format
-    try {
-      jwt.verify(refreshToken, REFRESH_SECRET);
-    } catch (error) {
-      // Always return 200 for tests
-      if (isTestMode) {
-        return res.status(200).json({ message: 'Logged out successfully (invalid token ignored)' });
-      }
-      return res.status(401).json({ error: 'Invalid refresh token format' });
-    }
-    
-    // Check if the refresh token exists in our store
-    if (!refreshTokens.has(refreshToken)) {
-      if (isTestMode) {
-        return res.status(200).json({ message: 'Logged out successfully (token not found)' });
-      }
-      return res.status(401).json({ error: 'Invalid or expired refresh token' });
-    }
-    
-    // Remove refresh token from store (if it exists)
-    refreshTokens.delete(refreshToken);
-    
-    res.json({ message: 'Logged out successfully' });
-  } catch (error) {
-    console.error('Logout error:', error);
-    res.status(500).json({ error: 'Logout failed' });
-  }
-});
-
-export default router;
