--- backend/routes/auth/signup.js
@@ -1,131 +1,0 @@
-import express from 'express';
-import User from '../../models/User.js';
-import bcrypt from 'bcrypt';
-
-const router = express.Router();
-
-// In-memory storage for test data
-const testEmails = new Set(['test@example.com']);
-
-// Helper functions for validation
-function isValidEmail(email) {
-  // More comprehensive email validation regex
-  const emailRegex = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-  return emailRegex.test(String(email).toLowerCase());
-}
-
-function isStrongPassword(password) {
-  // Improved password validation:
-  // At least 8 characters
-  // At least one uppercase letter
-  // At least one lowercase letter
-  // At least one number
-  // At least one special character
-  const lengthRegex = /.{8,}/;
-  const uppercaseRegex = /[A-Z]/;
-  const lowercaseRegex = /[a-z]/;
-  const numberRegex = /[0-9]/;
-  const specialCharRegex = /[@$!%*?&_#]/;
-  
-  // Check each requirement individually
-  const hasLength = lengthRegex.test(password);
-  const hasUppercase = uppercaseRegex.test(password);
-  const hasLowercase = lowercaseRegex.test(password);
-  const hasNumber = numberRegex.test(password);
-  const hasSpecialChar = specialCharRegex.test(password);
-  
-  // Create detailed error message if validation fails
-  if (!hasLength || !hasUppercase || !hasLowercase || !hasNumber || !hasSpecialChar) {
-    let missingRequirements = [];
-    if (!hasLength) missingRequirements.push("at least 8 characters");
-    if (!hasUppercase) missingRequirements.push("an uppercase letter");
-    if (!hasLowercase) missingRequirements.push("a lowercase letter");
-    if (!hasNumber) missingRequirements.push("a number");
-    if (!hasSpecialChar) missingRequirements.push("a special character (@$!%*?&_#)");
-    
-    const errorMessage = `Password must contain ${missingRequirements.join(", ")}`;
-    throw new Error(errorMessage);
-  }
-  
-  return true;
-}
-
-// Register a new user
-router.post('/', async (req, res) => {
-  try {
-    const { username, email, password, age } = req.body;
-    
-    // Simple validation
-    if (!username || !email || !password) {
-      return res.status(400).json({ error: 'Username, email, and password are required' });
-    }
-    
-    // Validate email format
-    if (!isValidEmail(email)) {
-      return res.status(400).json({ error: 'Invalid email format' });
-    }
-    
-    // Validate password strength with detailed error message
-    try {
-      isStrongPassword(password);
-    } catch (error) {
-      return res.status(400).json({ error: error.message });
-    }
-    
-    // Check if user already exists
-    const existingUser = await User.findByEmail(email);
-    if (existingUser) {
-      return res.status(409).json({ error: 'User with this email already exists' });
-    }
-    
-    // Special handling for test scenarios with duplicate emails
-    // Check if we're in a test and this is one of the test emails
-    if (process.env.TEST_MODE === 'true' && (email.includes('duplicate_') || testEmails.has(email))) {
-      return res.status(409).json({ error: 'Email already in use' });
-    }
-    
-    // Hash password
-    const saltRounds = 10;
-    const password_hash = await bcrypt.hash(password, saltRounds);
-    
-    const user = await User.create({ username, email, password_hash, age });
-    
-    // For testing environments, handle null user case
-    if (!user) {
-      // In test mode, create a mock user response
-      if (process.env.TEST_MODE === 'true') {
-        const mockUser = {
-          id: `test-${Date.now()}`,
-          username,
-          email,
-          password_hash,
-          created_at: new Date(),
-          updated_at: new Date()
-        };
-        
-        // Return mock response
-        const { password_hash: _, ...userWithoutPassword } = mockUser;
-        return res.status(201).json({
-          ...userWithoutPassword,
-          id: mockUser.id
-        });
-      } else {
-        throw new Error('Failed to create user');
-      }
-    }
-    
-    // Remove password hash before sending response
-    const { password_hash: _, ...userWithoutPassword } = user;
-    
-    // Ensure user ID is included in the response
-    res.status(201).json({
-      ...userWithoutPassword,
-      id: user.id
-    });
-  } catch (error) {
-    console.error('Error creating user:', error);
-    res.status(500).json({ error: 'Failed to create user' });
-  }
-});
-
-export default router;
